+++ src/obr/cli.py
@@ -272,6 +272,18 @@ def run(ctx: click.Context, **kwargs):
 def init(ctx: click.Context, **kwargs):
     # needs folder/.obr to exists before logger can be initialised
     ws_fold = kwargs.get("folder")
+
+    # Check if workspace already exists
+    if (Path(ws_fold) / "workspace").exists():
+        logger.warn(
+            f"Existing workspace found! Re-initializing over existing  workspace might"
+            f" lead to unexpected behaviour. Consider purging existing workspace with"
+            f" obr reset -w"
+        )
+        confirmed = click.confirm("Do you want to continue?", default=True)
+        if not confirmed:
+            return
+
     if ws_fold:
         (Path(ws_fold) / ".obr").mkdir(parents=True, exist_ok=True)
     else:
@@ -389,6 +401,17 @@ def query(ctx: click.Context, **kwargs):
     multiple=False,
     help="Path to script to apply to the workspace",
 )
+@click.option(
+    "--filter",
+    type=str,
+    multiple=True,
+    default=[],
+    help=(
+        "Pass a <key><predicate><value> value pair per occurrence of --filter."
+        " Predicates include ==, !=, <=, <, >=, >. For instance, obr submit --filter"
+        ' "solver==pisoFoam"'
+    ),
+)
 @click.pass_context
 def apply(ctx: click.Context, **kwargs):
     apply_file_path = Path(kwargs["file"]).resolve()
@@ -396,7 +419,8 @@ def apply(ctx: click.Context, **kwargs):
         logger.error(f"Could not find {kwargs['file']}")
         sys.exit(1)
 
-    project, jobs = cli_cmd_setup(kwargs)
+    global filtered_jobs
+    project, filtered_jobs = cli_cmd_setup(kwargs)
 
     os.environ["OBR_APPLY_FILE"] = str(apply_file_path)
     os.environ["OBR_APPLY_CAMPAIGN"] = kwargs.get("campaign", "")
@@ -557,12 +581,7 @@ def archive(ctx: click.Context, **kwargs):
     setup_logging()
 
     # setup project and jobs
-    project = OpenFOAMProject().init_project()
-    filters: list[str] = list(kwargs.get("filter", ()))
-    # check if given path points to valid project
-    if not is_valid_workspace(filters):
-        return
-    jobs = project.filter_jobs(filters, False)
+    project, jobs = cli_cmd_setup(kwargs)
 
     dry_run = kwargs.get("dry_run", False)
     branch_name = None