+++ CHANGELOG.rst
@@ -6,7 +6,8 @@ Changelog
 - Add grouping of jobs: `obr status --summarize N`, see https://github.com/hpsim/OBR/pull/200.
 
 0.3.0 (2024-03-03)
-------------------
+- Add warning if running obr init on existing workspace, see https://github.com/hpsim/OBR/pull/207
+- Fix issues with filters, see https://github.com/hpsim/OBR/pull/204
 - Add common notation see https://github.com/hpsim/OBR/pull/146
 - Make numberOfSubdomains argument in yaml consistent see https://github.com/hpsim/OBR/pull/148
 - Fix missing of groups in `obr run --list-operations` view, see https://github.com/hpsim/OBR/pull/159.